requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
frozenlist==1.4.1
idna==3.7
inflection==0.5.1
multidict==6.0.5
packaging==24.1
pillow==10.3.0
psycopg2==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==70.1.1
sqlparse==0.5.0
twilio==9.2.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.4
